Add created_at and updated_at fields. Move the admin info to be under general directory. Rerun migration with the new fields
